frontend/requirements.txt
flask>=3.0.0
pandas>=2.0.0
numpy>=1.26.0
duckdb>=1.0.0
google-generativeai>=0.7.0
